requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.7
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tornado==6.0.4
urllib3==1.25.9
